commit fix view admin

+++ app/Http/Controllers/BannerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Images;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class BannerController extends Controller
@@ -32,61 +33,63 @@ class BannerController extends Controller
      */
     public function store(Request $request, Images $images)
     {
-        $requestData = $request->all();
-        if ($request->hasFile('image')) {
-            $image = $request->file('image');
-            $extension = $image->getClientOriginalExtension();
-            $pathFile = $image->storeAs('storage/banners', $image->getClientOriginalName(), 'public');
-            $image->move(\public_path("storage/banners/"), $image->getClientOriginalName());
-            Images::query()->create(
-                [
-                    'image' => $pathFile,
-                    'is_active' => 0,
-                    'type' => $extension,
-                    'created_at' => $requestData['create_at'],
-                    'updated_at' => $requestData['updated_at'],
-                ]
-            );
-            return redirect()->route('banner.index');
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $requestData = $request->all();
+            if ($request->hasFile('image')) {
+                $image = $request->file('image');
+                $extension = $image->getClientOriginalExtension();
+                $pathFile = $image->storeAs('storage/banners', $image->getClientOriginalName(), 'public');
+                $image->move(\public_path("storage/banners/"), $image->getClientOriginalName());
+                Images::query()->create(
+                    [
+                        'image' => $pathFile,
+                        'is_active' => 0,
+                        'type' => $extension,
+                        'created_at' => $requestData['create_at'],
+                        'updated_at' => $requestData['updated_at'],
+                    ]
+                );
+                return redirect()->route('banner.index');
+            } else {
+                return redirect()->route('banner.create');
+            }
         } else {
-            return redirect()->route('banner.create');
+            return redirect()->route('login');
         }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Images $image)
-    {
-        //
-    }
+
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Images $image)
-    {
-        echo 'check';
-    }
+
 
     /**
      * Update the specified resource in storage.
      */
     public function update($id)
     {
-        $active = Images::query()->where('id', $id)->value('is_active');
-        if ($active == 1) {
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $active = Images::query()->where('id', $id)->value('is_active');
+            if ($active == 1) {
 
-            Images::query()->where('id', $id)->update([
-                'is_active' => 0
-            ]);
+                Images::query()->where('id', $id)->update([
+                    'is_active' => 0
+                ]);
+            } else {
+
+                Images::query()->where('id', $id)->update([
+                    'is_active' => 1
+                ]);
+            }
+            return redirect()->route('banner.index');
         } else {
-
-            Images::query()->where('id', $id)->update([
-                'is_active' => 1
-            ]);
+            return redirect()->route('login');
         }
-        return redirect()->route('banner.index');
     }
 
 
@@ -95,7 +98,11 @@ class BannerController extends Controller
      */
     public function destroy($image)
     {
-        Images::query()->where('id', $image)->delete();
-        return redirect()->route('banner.index');
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            Images::query()->where('id', $image)->delete();
+            return redirect()->route('banner.index');
+        } else {
+            return redirect()->route('login');
+        }
     }
 }
